test: mark crypto-dependent tests
